Final commit for the version 1.8.16
Removing the calls to require() in the code. Stressing again that
this is the very last version of the package. Removing the incorrect
URL for the kernelhr package. Removing "still under research" for
modpartltraj. Accounting for the warning concerning the DESCRIPTION
file (date format, period at the end of sentence, upper case for the title,
etc.).